requirements.txt
asgiref==3.11.0
Django==6.0
django-enumfield==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2
